+++ tests/testplotfit.R
@@ -29,5 +29,18 @@ if (visualize)
   plotfit2pdf(f , items = 6, BMDoutput = b,  BMDtype = "xfold") 
   plotfit2pdf(f , items = c("12.1", "4", "70"),  BMDoutput = r, BMDtype = "zSD") 
   
+  
+  # Metabolo ex. case where there is a pb of items ordering 
+  datafilename <- system.file("extdata", "metabolo_sample.txt", package="DRomics")
+  o <- continuousomicdata(datafilename, backgrounddose = 0, check = TRUE)
+  s <- itemselect(o, select.method = 'quadratic', FDR = 0.05)
+  f <- drcfit(s, progressbar = FALSE, parallel = 'no')
+  plot(f, plot.type = 'dose_fitted', dose_log_transfo = TRUE)
+  r <- bmdcalc(f, z = 1, x = 10)
+  plot(f, plot.type = 'dose_fitted', BMDoutput = r, BMDtype = 'zSD', dose_log_transfo = TRUE) # good order
+  plotfit2pdf(f) # good order - initial order by p-value of the selection
+  plotfit2pdf(f , BMDoutput = r, BMDtype = "zSD")  # alphabetic order WHY ???????????????
+  b <- bmdboot(r, niter = 10)
+  plotfit2pdf(f , BMDoutput = b, BMDtype = "zSD") # alphabetic order
 }